BAEL-1862 Move the Junit 5 logic in the right module
-Renamed *Test to *UnitTest
