+++ testing-modules/junit-5/src/test/java/com/baeldung/methodorders/NameAscendingOrderOfExecutionUnitTest.java
@@ -7,8 +7,8 @@ import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
-@FixMethodOrder(MethodSorters.DEFAULT)
-public class DefaultOrderOfExecutionTest {
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class NameAscendingOrderOfExecutionUnitTest {
     private static StringBuilder output = new StringBuilder("");
 
     @Test
@@ -28,6 +28,6 @@ public class DefaultOrderOfExecutionTest {
 
     @AfterClass
     public static void assertOutput() {
-        assertEquals(output.toString(), "cab");
+        assertEquals(output.toString(), "abc");
     }
 }
